+++ R/facets.R
@@ -6,6 +6,15 @@
 #' @param rows,cols Number of rows and columns.
 #' 
 #' @details Each serie, i.e.:\code{\link{e_bar}} will be plotted against a facet.
+#' 
+#' @examples 
+#' mtcars %>% 
+#'   e_charts(mpg) %>% 
+#'   e_scatter(qsec, mpg) %>% 
+#'   e_line(wt, gridIndex = 1) %>% 
+#'   e_facet(2)
+#' 
+#' @keywords interval
 e_facet <- function(e, rows = 1, cols = 1){
   
   # number of series